Allow gltf extension factories to be async (to enable dynamic imports)
