+++ packages/dev/loaders/src/SPLAT/splatFileLoader.dynamic.ts
