+++ packages/dev/loaders/src/STL/stlFileLoader.dynamic.ts
