+++ packages/dev/loaders/src/dynamic.ts
@@ -1,12 +1,14 @@
-import { registerOBJLoader } from "./OBJ/objDynamicFileLoader";
-import { registerGLTFLoader } from "./glTF/glTFDynamicFileLoader";
-import { registerSPLATLoader } from "./SPLAT/splatDynamicFileLoader";
-import { registerSTLLoader } from "./STL/stlDynamicFileLoader";
+import { registerOBJLoader } from "./OBJ/objFileLoader.dynamic";
+import { registerGLTFLoader } from "./glTF/glTFFileLoader.dynamic";
+import { registerDefaultGLTFExtensions } from "./glTF/2.0/Extensions/dynamic";
+import { registerSPLATLoader } from "./SPLAT/splatFileLoader.dynamic";
+import { registerSTLLoader } from "./STL/stlFileLoader.dynamic";
 
 //test
-export function registerLoaders() {
+export function registerDefaultLoaders() {
     registerOBJLoader();
     registerGLTFLoader();
+    registerDefaultGLTFExtensions();
     registerSPLATLoader();
     registerSTLLoader();
 }
