+++ packages/dev/loaders/src/glTF/2.0/Extensions/dynamic.ts
@@ -1,0 +1,5 @@
+import { registerKHR_materials_pbrSpecularGlossiness } from "./KHR_materials_pbrSpecularGlossiness.dynamic";
+
+export function registerDefaultGLTFExtensions() {
+    registerKHR_materials_pbrSpecularGlossiness();
+}
