+++ packages/dev/loaders/src/glTF/2.0/glTFLoader.ts
@@ -76,6 +76,8 @@ import type { AssetContainer } from "core/assetContainer";
 import type { AnimationPropertyInfo } from "./glTFLoaderAnimation";
 import { nodeAnimationData } from "./glTFLoaderAnimation";
 import type { IObjectInfo } from "core/ObjectModel/objectModelInterfaces";
+import { GLTFLoaderExtensionRegistry } from "./glTFLoaderExtensionRegistry";
+import type { ExtensionFactory } from "./glTFLoaderExtensionRegistry";
 export { GLTFFileLoader };
 
 interface TypedArrayLike extends ArrayBufferView {
@@ -92,10 +94,6 @@ interface ILoaderProperty extends IProperty {
     _activeLoaderExtensionFunctions: {
         [id: string]: boolean;
     };
-}
-
-interface IRegisteredExtension {
-    factory: (loader: GLTFLoader) => IGLTFLoaderExtension;
 }
 
 interface IWithMetadata {
@@ -212,8 +210,6 @@ export class GLTFLoader implements IGLTFLoader {
     private _defaultBabylonMaterialData: { [drawMode: number]: Material } = {};
     private readonly _postSceneLoadActions = new Array<() => void>();
 
-    private static _RegisteredExtensions: { [name: string]: IRegisteredExtension } = {};
-
     /**
      * The default glTF sampler.
      */
@@ -224,14 +220,8 @@ export class GLTFLoader implements IGLTFLoader {
      * @param name The name of the loader extension.
      * @param factory The factory function that creates the loader extension.
      */
-    public static RegisterExtension(name: string, factory: (loader: GLTFLoader) => IGLTFLoaderExtension): void {
-        if (GLTFLoader.UnregisterExtension(name)) {
-            Logger.Warn(`Extension with the name '${name}' already exists`);
-        }
-
-        GLTFLoader._RegisteredExtensions[name] = {
-            factory: factory,
-        };
+    public static RegisterExtension(name: string, factory: ExtensionFactory): void {
+        GLTFLoaderExtensionRegistry.Register(name, false, factory);
     }
 
     /**
@@ -240,12 +230,7 @@ export class GLTFLoader implements IGLTFLoader {
      * @returns A boolean indicating whether the extension has been unregistered
      */
     public static UnregisterExtension(name: string): boolean {
-        if (!GLTFLoader._RegisteredExtensions[name]) {
-            return false;
-        }
-
-        delete GLTFLoader._RegisteredExtensions[name];
-        return true;
+        return GLTFLoaderExtensionRegistry.Unregister(name);
     }
 
     /**
@@ -396,14 +381,13 @@ export class GLTFLoader implements IGLTFLoader {
 
     private _loadAsync<T>(rootUrl: string, fileName: string, nodes: Nullable<Array<number>>, resultFunc: () => T): Promise<T> {
         return Promise.resolve()
-            .then(() => {
+            .then(async () => {
                 this._rootUrl = rootUrl;
                 this._uniqueRootUrl = !rootUrl.startsWith("file:") && fileName ? rootUrl : `${rootUrl}${Date.now()}/`;
                 this._fileName = fileName;
                 this._allMaterialsDirtyRequired = false;
 
-                this._loadExtensions();
-                this._checkExtensions();
+                await this._loadExtensionsAsync();
 
                 const loadingToReadyCounterName = `${GLTFLoaderState[GLTFLoaderState.LOADING]} => ${GLTFLoaderState[GLTFLoaderState.READY]}`;
                 const loadingToCompleteCounterName = `${GLTFLoaderState[GLTFLoaderState.LOADING]} => ${GLTFLoaderState[GLTFLoaderState.COMPLETE]}`;
@@ -564,30 +548,38 @@ export class GLTFLoader implements IGLTFLoader {
         }
     }
 
-    private _loadExtensions(): void {
-        for (const name in GLTFLoader._RegisteredExtensions) {
+    private async _loadExtensionsAsync() {
+        const extensionPromises: Promise<IGLTFLoaderExtension>[] = [];
+
+        for (const [name, registeredExtension] of GLTFLoaderExtensionRegistry.RegisteredExtensions) {
             // Don't load explicitly disabled extensions.
             if (this.parent.extensionOptions[name]?.enabled === false) {
                 // But warn if the disabled extension is used by the model.
-                if (this.isExtensionUsed(name)) {
+                if (registeredExtension.isGLTFExtension && this.isExtensionUsed(name)) {
                     Logger.Warn(`Extension ${name} is used but has been explicitly disabled.`);
                 }
-            } else {
-                const extension = GLTFLoader._RegisteredExtensions[name].factory(this);
-                if (extension.name !== name) {
-                    Logger.Warn(`The name of the glTF loader extension instance does not match the registered name: ${extension.name} !== ${name}`);
-                }
+            }
+            // Load loader extensions that are not a glTF extension, as well as extensions that are glTF extensions and are used by the model.
+            else if (!registeredExtension.isGLTFExtension || this.isExtensionUsed(name)) {
+                extensionPromises.push(
+                    (async () => {
+                        const extension = await registeredExtension.factory(this);
+                        if (extension.name !== name) {
+                            Logger.Warn(`The name of the glTF loader extension instance does not match the registered name: ${extension.name} !== ${name}`);
+                        }
 
-                this._extensions.push(extension);
-                this._parent.onExtensionLoadedObservable.notifyObservers(extension);
+                        this._parent.onExtensionLoadedObservable.notifyObservers(extension);
+                        return extension;
+                    })()
+                );
             }
         }
 
+        this._extensions.push(...(await Promise.all(extensionPromises)));
+
         this._extensions.sort((a, b) => (a.order || Number.MAX_VALUE) - (b.order || Number.MAX_VALUE));
         this._parent.onExtensionLoadedObservable.clear();
-    }
 
-    private _checkExtensions(): void {
         if (this._gltf.extensionsRequired) {
             for (const name of this._gltf.extensionsRequired) {
                 const available = this._extensions.some((extension) => extension.name === name && extension.enabled);
